requirements.txt
adium-theme-ubuntu==0.3.4
argcomplete==1.9.4
astroid==1.6.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.5.2
cfscrape==1.9.0
chardet==3.0.4
Cheetah==2.4.4
click==6.7
configobj==5.0.6
configparser==3.5.0
crcmod==1.7
cryptography==1.2.3
cycler==0.10.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
gyp==0.1
html5lib==0.999
idna==2.6
ipaddress==1.0.16
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
nltk==3.2.5
numpy==1.13.3
pdfminer==20140328
Pillow==5.1.0
ply==3.7
poster==0.8.1
pyasn1==0.1.9
PyBluez==0.18
pycparser==2.14
pycurl==7.43.0
PyExecJS==1.4.1
pygobject==3.20.0
pylint==1.9.3
pyOpenSSL==0.15.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-docx==0.8.6
python-pptx==0.6.9
pytz==2017.3
requests==2.18.4
selenium==3.6.0
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
textract==0.4.0
trimage==1.0.2
unity-lens-photos==1.0
urllib3==1.22
variety==0.6.5
variety-slideshow==0.1
Werkzeug==0.14.1
wrapt==1.10.11
xattr==0.7.8
XlsxWriter==1.0.5
